Update event query to pull active events
Updates event query to pull events with a start date greater than
yesterday, recurring events without an end date, and recurring events
with an end date greater than yesterday. This query retrieves all active
events as of the day before. The reason why it pulls in events from
current time minus one day is because an event that occured a few hours
ago may still be in progress. The returned query results are handed off
to the scheduler to determine actual upcoming events based on any
current event instances and upcoming schedule.

This is needed, because the previous query was pulling in inactive
events causing a bottle neck in the ice cube gem that handles all the
scheduling which calculates if there are events in progress and/or when
the next event if any would occur.